+++ VDA-ERC20/ChangeLog.md
@@ -1,3 +1,7 @@
+2023-04-06 (v0.2.0)
+-------------------
+- Added scripts for multi-sign support
+
 2023-04-03 (v0.1.0)
 -------------------
 Updates on audit report